java-baseball-6-main/src/main/java/baseball/domain/ComputerNumber.java
package baseball.domain;

import camp.nextstep.edu.missionutils.Randoms;
import java.util.ArrayList;
import java.util.List;

public class ComputerNumber {
    public List<Integer> generate() {
        List<Integer> computerNumber = new ArrayList<>();

        while (computerNumber.size() < 3) {
            Integer randomNumber = Randoms.pickNumberInRange(1, 9);

            if (!computerNumber.contains(randomNumber)) {
                computerNumber.add(randomNumber);
            }
        }

        return computerNumber;
    }
}
